mangaged to draw le grid but unfortunatelly I need to slightly optimize it
mangaged to draw le grid but unfortunatelly I need to slightly optimize it

smol changes

+++ src/cli.zig
@@ -4,7 +4,7 @@ const builtin = @import("builtin");
 const Seto = @import("main.zig").Seto;
 const Function = @import("config.zig").Function;
 
-fn hexToRgba(hex: ?[]const u8) ![4]f64 {
+fn hexToRgba(hex: ?[]const u8) ![4]f32 {
     if (hex == null) {
         return error.ArgumentMissing;
     }
@@ -15,10 +15,10 @@ fn hexToRgba(hex: ?[]const u8) ![4]f64 {
         return error.InvalidColor;
     }
 
-    const r: f64 = @floatFromInt(try std.fmt.parseInt(u8, hex.?[0 + start .. 2 + start], 16));
-    const g: f64 = @floatFromInt(try std.fmt.parseInt(u8, hex.?[2 + start .. 4 + start], 16));
-    const b: f64 = @floatFromInt(try std.fmt.parseInt(u8, hex.?[4 + start .. 6 + start], 16));
-    const a: f64 = if (hex.?.len > 6 + start) @floatFromInt(try std.fmt.parseInt(u8, hex.?[6 + start .. 8 + start], 16)) else 255;
+    const r: f32 = @floatFromInt(try std.fmt.parseInt(u8, hex.?[0 + start .. 2 + start], 16));
+    const g: f32 = @floatFromInt(try std.fmt.parseInt(u8, hex.?[2 + start .. 4 + start], 16));
+    const b: f32 = @floatFromInt(try std.fmt.parseInt(u8, hex.?[4 + start .. 6 + start], 16));
+    const a: f32 = if (hex.?.len > 6 + start) @floatFromInt(try std.fmt.parseInt(u8, hex.?[6 + start .. 8 + start], 16)) else 255;
 
     return .{
         r / 255,
@@ -88,7 +88,7 @@ pub fn parseArgs(seto: *Seto) void {
             seto.config.grid.offset = parseIntArray(getNextArg(&args, arg), ",") catch printAndExit("Incorrect argument for \"{s}\"\n", arg);
         } else if (std.mem.eql(u8, arg, "--line-width")) {
             const line_width = getNextArg(&args, arg);
-            seto.config.grid.line_width = std.fmt.parseFloat(f64, line_width) catch printAndExit("Incorrect argument for \"{s}\"\n", arg);
+            seto.config.grid.line_width = std.fmt.parseFloat(f32, line_width) catch printAndExit("Incorrect argument for \"{s}\"\n", arg);
         } else if (std.mem.eql(u8, arg, "--selected-line-width")) {
             const line_width = getNextArg(&args, arg);
             seto.config.grid.selected_line_width = std.fmt.parseFloat(f64, line_width) catch printAndExit("Incorrect argument for --line-width {s}\n", line_width);
